@@ -5,17 +5,17 @@ repository: https://github.com/josiahsrc/flutter_dough
 homepage: https://josiahsaunders.com/
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  sensors: ">=0.4.2+2 <2.0.0"
-  equatable: ^1.2.5
-  vector_math: ^2.0.8
-  provider: ^4.3.2
+  sensors: ^2.0.0
+  equatable: ^2.0.0
+  vector_math: ^2.1.0
+  provider: ^5.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  very_good_analysis: ^1.0.0
+  very_good_analysis: ^2.0.0
