@@ -26,7 +26,6 @@ class DoughController with ChangeNotifier {
   DoughController() : super();
 
   final _statusListeners = ObserverList<DoughStatusCallback>();
-
   bool _isActive = false;
   Offset _origin = Offset.zero;
   Offset _target = Offset.zero;
@@ -68,8 +67,8 @@ class DoughController with ChangeNotifier {
   ///
   /// **A squish can't already be active when calling this function.**
   void start({
-    Offset origin,
-    Offset target,
+    Offset? origin,
+    Offset? target,
   }) {
     assert(!isActive);
 
@@ -89,8 +88,8 @@ class DoughController with ChangeNotifier {
   ///
   /// **A squish must already be active when calling this function.**
   void update({
-    Offset origin,
-    Offset target,
+    Offset? origin,
+    Offset? target,
   }) {
     assert(isActive);
 
@@ -117,7 +116,7 @@ class DoughController with ChangeNotifier {
 
   void _notifyStatusListeners(DoughStatus status) {
     for (final listener in _statusListeners) {
-      listener?.call(status);
+      listener.call(status);
     }
   }
 }
